+++ R/SelectBoost-package.R
@@ -13,7 +13,7 @@
 #' @docType package
 #' @name SelectBoost
 #' @references \emph{selectBoost: a general algorithm to enhance the performance of variable selection methods in correlated datasets}, Ismaïl Aouadi, Nicolas Jung, Raphael Carapito, Laurent Vallat, Seiamak Bahram, Myriam Maumy-Bertrand, Frédéric Bertrand, \url{https://arxiv.org/abs/1810.01670}
-# @import lars glmnet igraph parallel msgps Rfast doMC foreach
+# @import lars glmnet igraph parallel msgps Rfast doParallel foreach
 #' @import lars glmnet msgps
 #' @importFrom methods new
 #' @importFrom utils head tail
@@ -22,8 +22,8 @@
 #' @importFrom graphics abline axis barplot matplot mtext par plot points segments
 #' @importFrom stats binom.test median plogis pt qbinom quantile rbinom rnorm runif sd t.test wilcox.test
 #' @importFrom Rfast vmf.mle rvmf
-#' @importFrom parallel detectCores makeCluster clusterEvalQ clusterExport parLapply stopCluster
-#' @importFrom doMC registerDoMC
+#' @importFrom parallel detectCores makeCluster makeForkCluster makePSOCKcluster clusterEvalQ clusterExport parLapply stopCluster
+#' @importFrom doParallel registerDoParallel
 #' @importFrom foreach foreach "%dopar%"
 #' @importFrom igraph graph.adjacency infomap.community communities
 #' @importFrom grDevices rgb gray